+++ app/handler/api.go
@@ -1,9 +1,7 @@
 package handler
 
 import (
-	"github.com/dingoblog/dingo/app/model"
 	"github.com/dinever/golf"
-	"strconv"
 )
 
 // APIDocumentationHandler shows which routes match with what functionality,
@@ -29,170 +27,6 @@ func APIDocumentationHandler(ctx *golf.Context) {
 		"user_email_url":        "/api/users/email/:email",
 	}
 	ctx.JSONIndent(routes, "", "  ")
-}
-
-// APICommentsHandler retrieves all the comments.
-func APICommentsHandler(ctx *golf.Context) {
-	ctx.JSONIndent(map[string]interface{}{
-		"message": "Not implemented",
-	}, "", "  ")
-}
-
-// APICommentHandler retrieves a comment with the given comment id.
-func APICommentHandler(ctx *golf.Context) {
-	id, err := strconv.Atoi(ctx.Param("id"))
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	comment := &model.Comment{Id: int64(id)}
-	err = comment.GetCommentById()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(comment, "", "  ")
-}
-
-// APICommentPostHandler retrives the tag with the given post id.
-func APICommentPostHandler(ctx *golf.Context) {
-	id, err := strconv.Atoi(ctx.Param("id"))
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	comments := new(model.Comments)
-	err = comments.GetCommentsByPostId(int64(id))
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(comments, "", "  ")
-}
-
-// APIPostsHandler gets every page, ordered by publication date.
-func APIPostsHandler(ctx *golf.Context) {
-	posts := new(model.Posts)
-	err := posts.GetAllPostList(false, true, "published_at DESC")
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(posts, "", "  ")
-}
-
-// APIPostHandler retrieves the post with the given ID.
-func APIPostHandler(ctx *golf.Context) {
-	id, err := strconv.Atoi(ctx.Param("id"))
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	post := &model.Post{Id: int64(id)}
-	err = post.GetPostById()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(post, "", "  ")
-}
-
-// APIPostSlugHandler retrieves the post with the given slug.
-func APIPostSlugHandler(ctx *golf.Context) {
-	slug := ctx.Param("slug")
-	post := new(model.Post)
-	err := post.GetPostBySlug(slug)
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(post, "", "  ")
-}
-
-// APITagsHandler retrieves all the tags.
-func APITagsHandler(ctx *golf.Context) {
-	tags := new(model.Tags)
-	err := tags.GetAllTags()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(tags, "", "  ")
-}
-
-// APITagHandler retrieves the tag with the given id.
-func APITagHandler(ctx *golf.Context) {
-	id, err := strconv.Atoi(ctx.Param("id"))
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	tag := &model.Tag{Id: int64(id)}
-	err = tag.GetTag()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(tag, "", "  ")
-}
-
-// APITagSlugHandler retrieves the tag(s) with the given slug.
-func APITagSlugHandler(ctx *golf.Context) {
-	slug := ctx.Param("slug")
-	tags := &model.Tag{Slug: slug}
-	err := tags.GetTagBySlug()
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	ctx.JSONIndent(tags, "", "  ")
-}
-
-// APIUsersHandler retrieves all users.
-func APIUsersHandler(ctx *golf.Context) {
-	ctx.JSONIndent(map[string]interface{}{
-		"message": "Not implemented",
-	}, "", "  ")
-}
-
-// APIUserHandler retrieves the user with the given id.
-func APIUserHandler(ctx *golf.Context) {
-	id, err := strconv.Atoi(ctx.Param("id"))
-	if err != nil {
-		handleErr(ctx, 500, err)
-		return
-	}
-	user := &model.User{Id: int64(id)}
-	err = user.GetUserById()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(user, "", "  ")
-}
-
-// APIUserSlugHandler retrives the user with the given slug.
-func APIUserSlugHandler(ctx *golf.Context) {
-	slug := ctx.Param("slug")
-	user := &model.User{Slug: slug}
-	err := user.GetUserBySlug()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(user, "", "  ")
-}
-
-// APIUserEmailHandler retrieves the user with the given email.
-func APIUserEmailHandler(ctx *golf.Context) {
-	email := ctx.Param("email")
-	user := &model.User{Email: email}
-	err := user.GetUserByEmail()
-	if err != nil {
-		handleErr(ctx, 404, err)
-		return
-	}
-	ctx.JSONIndent(user, "", "  ")
 }
 
 // handleErr sends the staus code and error message formatted as JSON.
